+++ src/main/scala/com/abc/Transaction.scala
@@ -2,7 +2,7 @@ package com.abc
 
 import java.time.Instant
 
-case class Transaction(val amount: Double) {
+case class Transaction(val amount: BigDecimal) {
   val transactionDate = Instant.now
 }
